Apply same logic to x axis
